Move stream files to the separate directory.
