+++ Library/GCC/crt0.h
@@ -3,7 +3,7 @@
 
 #include "../stdlib.h"
 #include "../stdio.h"
-#include "../Streams/streams.h"
+#include "../stdio/streams/streams.h"
 
 void _start(char *path, char *parameters);
 char **parse_parameters(char *path, char *parameters, int *count);
