+++ Library/stdio/streams/console_stream.c
