+++ Library/stdio/streams/console_stream.h
@@ -2,8 +2,8 @@
 #define CONSOLE_STREAM_H
 
 #include <stdint.h>
-#include "../stdio.h"
-#include "../micros.h"
+#include "../../stdio.h"
+#include "../../micros.h"
 
 FILE *streams_set_stream_as_console(FILE *stream);
 int streams_console_fetch(FILE *stream);
