+++ Library/stdio/streams/keyboard_stream.c
